ONE VISA CONNECT/Customer/Android App/my app/backend/Auth API/requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

flask
flask-jwt-extended
flask-marshmallow
flask-restful
flask-sqlalchemy
marshmallow-sqlalchemy
marshmallow
cryptography
urllib3==1.25.9
flask-cors
cryptography
PyMySQL
python-dotenv
requests
pycryptodomex
pycryptodome
twilio
